edgeServiceLib/device-management/redis-recv/dynamicLib/requirements.txt
flask
requests
kafka-python